Remove frozen evaluator stack
PARITY.md is at 57/0/0/3; vix-daemon was already on the machine and vix-wire now runs through the machine fleet seam. Relocate the surviving lua parser/highlight coverage, then delete oracle, engine, graph, their evaluator suites, and orphaned value/fetch compatibility glue.
